backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
psycopg2-binary>=2.9
pydantic>=2.6
pydantic-settings>=2.2
PyJWT>=2.8
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
